# Index eager load performance