--- a/lib/administrate/base_dashboard.rb
+++ b/lib/administrate/base_dashboard.rb
@@ -97,6 +97,10 @@ module Administrate
       attribute_includes(show_page_attributes)
     end
 
+    def item_associations
+      attribute_associated(show_page_attributes)
+    end
+
     private
 
     def attribute_not_found_message(attr)
@@ -104,6 +108,14 @@ module Administrate
     end
 
     def attribute_includes(attributes)
+      attributes.map do |key|
+        field = attribute_type_for(key)
+
+        key if field.eager_load?
+      end.compact
+    end
+
+    def attribute_associated(attributes)
       attributes.map do |key|
         field = attribute_type_for(key)