--- a/lib/administrate/field/belongs_to.rb
+++ b/lib/administrate/field/belongs_to.rb
@@ -13,6 +13,10 @@ module Administrate
         end
       end
 
+      def self.eager_load?
+        true
+      end
+
       def permitted_attribute
         foreign_key
       end